Moving all annotations to Java API
