+++ stratosphere-addons/array-datamodel/src/main/java/eu/stratosphere/arraymodel/example/WordCountArrayTuples.java
@@ -22,9 +22,9 @@ import eu.stratosphere.api.common.operators.FileDataSink;
 import eu.stratosphere.api.common.operators.FileDataSource;
 import eu.stratosphere.api.common.operators.base.MapOperatorBase;
 import eu.stratosphere.api.common.operators.base.ReduceOperatorBase;
-import eu.stratosphere.api.common.operators.base.ReduceOperatorBase.Combinable;
 import eu.stratosphere.api.java.record.functions.FunctionAnnotation.ConstantFields;
 import eu.stratosphere.api.java.record.io.TextInputFormat;
+import eu.stratosphere.api.java.record.operators.ReduceOperator.Combinable;
 import eu.stratosphere.arraymodel.ArrayModelPlan;
 import eu.stratosphere.arraymodel.functions.DataTypes;
 import eu.stratosphere.arraymodel.functions.MapFunction;
