+++ stratosphere-core/src/main/java/eu/stratosphere/api/common/operators/base/CoGroupOperatorBase.java
@@ -13,11 +13,6 @@
 
 package eu.stratosphere.api.common.operators.base;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
 import eu.stratosphere.api.common.functions.GenericCoGrouper;
 import eu.stratosphere.api.common.operators.DualInputOperator;
 import eu.stratosphere.api.common.operators.util.UserCodeClassWrapper;
@@ -36,8 +31,15 @@ import eu.stratosphere.api.common.operators.util.UserCodeWrapper;
  */
 public class CoGroupOperatorBase<T extends GenericCoGrouper<?, ?, ?>> extends DualInputOperator<T> {
 	
+	private boolean combinableFirst;
+
+	private boolean combinableSecond;
+
+	
 	public CoGroupOperatorBase(UserCodeWrapper<T> udf, int[] keyPositions1, int[] keyPositions2, String name) {
 		super(udf, keyPositions1, keyPositions2, name);
+		this.combinableFirst = false;
+		this.combinableSecond = false;
 	}
 	
 	public CoGroupOperatorBase(T udf, int[] keyPositions1, int[] keyPositions2, String name) {
@@ -47,20 +49,21 @@ public class CoGroupOperatorBase<T extends GenericCoGrouper<?, ?, ?>> extends Du
 	public CoGroupOperatorBase(Class<? extends T> udf, int[] keyPositions1, int[] keyPositions2, String name) {
 		this(new UserCodeClassWrapper<T>(udf), keyPositions1, keyPositions2, name);
 	}
-
+	
 	public boolean isCombinableFirst() {
-		return getUserCodeAnnotation(CombinableFirst.class) != null;
+		return this.combinableFirst;
+	}
+	
+	public void setCombinableFirst(boolean combinableFirst) {
+		this.combinableFirst = combinableFirst;
 	}
 	
 	public boolean isCombinableSecond() {
-		return getUserCodeAnnotation(CombinableSecond.class) != null;
+		return this.combinableSecond;
 	}
 
-	@Retention(RetentionPolicy.RUNTIME)
-	@Target(ElementType.TYPE)
-	public static @interface CombinableFirst {};
-	
-	@Retention(RetentionPolicy.RUNTIME)
-	@Target(ElementType.TYPE)
-	public static @interface CombinableSecond {};
+	public void setCombinableSecond(boolean combinableSecond) {
+		this.combinableSecond = combinableSecond;
+	}
+
 }
